Updated Commands MID
From what I’ve seen in the traffic, Command 0x21 has a single byte sub command (arguments index 0) with values of only 0x41 or 0x43, with the remainder of the arguments looking like chars and control codes akin to those sent to the display driver. I _think_ this is to draw the MID display, with the sub command being one of the rows. I’ve not used a MID _ever_.. so I’m not sure, but the absence of this command or altering it does not seem to affect the other displays, so I’ll roll with it for now.